Concern 0-C-pre: fix Swift 6.2.3 type-checker failures (Xcode 26.3) (#3)
* Fix Swift 6.2.3 type-checker failures under Xcode 26.3

These are Swift 6.2.3 (Xcode 26.3) inference failures, not deployment-target
issues — they reproduce at 26.4, 26.0, and 18.0. Xcode 26.4 / Swift 6.3
resolves them, which is why upstream does not see them. Cherry-pickable as-is.


* Prefer Int64(exactly:)! over truncatingIfNeeded for ObjectBox shadow

ObjectBox's UntypedIdBase extension on Swift.Int64 shadows Int64(_:) for
UInt64 under MemberImportVisibility; init(exactly:) does not collide and
preserves trapping. Document as §7.12 so the lesson is not re-earned.


---------

+++ ColorPickerView.swift
@@ -90,7 +90,7 @@ struct AccentColorPickerView: View {
             .gesture(
                 DragGesture(minimumDistance: 0).onChanged { value in
                     saturation = max(0, min(1, value.location.x / size))
-                    brightness = max(0, min(1, 1 - value.location.y / size))
+                    brightness = max(0, min(1, 1 - Double(value.location.y / size)))
                 }
             )
         }
